+++ pumpkin-solver/src/engine/cp/trailed/mod.rs
@@ -1,6 +1,6 @@
-mod trailed_assignments;
 mod trailed_change;
 mod trailed_integer;
-pub(crate) use trailed_assignments::TrailedValues;
+mod trailed_values;
 pub(crate) use trailed_change::TrailedChange;
 pub(crate) use trailed_integer::TrailedInteger;
+pub(crate) use trailed_values::TrailedValues;
