+++ pumpkin-solver/src/engine/cp/trailed/trailed_values.rs
